Add real-time dashboard and Axum ingestion API code generation
This commit implements two major features for nomnom:
1. Axum-based ingestion API server generation
2. Real-time dashboard generation (Axum backend + React frontend)

## Ingestion API Server (Axum)

New CLI command: `nomnom generate-ingestion-api`

Generates a production-ready Axum REST API server for data ingestion with:
- Auto-generated endpoints from entity YAML definitions
- Type-safe request/response handling with custom parsers
- Diesel ORM integration for database operations
- Streaming ingestion support (newline-delimited JSON)
- Comprehensive error handling and validation
- Database connection pooling and migration support
- Docker deployment configuration

Code generation modules:
- src/codegen/ingestion_server/main_rs.rs - Axum server setup
- src/codegen/ingestion_server/handlers_rs.rs - REST endpoint handlers
- src/codegen/ingestion_server/parsers_rs.rs - Custom type parsers
- src/codegen/ingestion_server/models_rs.rs - Request/response types
- src/codegen/ingestion_server/database_rs.rs - Database operations
- src/codegen/ingestion_server/error_rs.rs - Error handling
- src/codegen/ingestion_server/cargo_toml.rs - Cargo.toml generation

## Real-Time Dashboard

New CLI command: `nomnom generate-dashboard`

Generates a full-stack real-time monitoring dashboard with:
- Axum WebSocket backend for real-time data streaming
- React + Vite frontend with Tailwind CSS
- Polling-based architecture (no database triggers needed)
- Initial data loading when clients connect
- Auto-discovery of all persistent entities
- Database-agnostic (PostgreSQL, MySQL, MariaDB support)
- Configurable polling intervals and record limits
- Docker deployment ready

Dashboard features:
- Live entity tables with configurable field display
- Connection status indicator
- Entity statistics (total record counts)
- REST API endpoints for metadata and health checks
- WebSocket for real-time updates

Code generation modules:
- src/codegen/dashboard/axum_backend.rs - Axum WebSocket server
- src/codegen/dashboard/react_frontend.rs - React UI components
- src/codegen/dashboard/fastapi_backend.rs - FastAPI backend (legacy)
- src/codegen/dashboard/sql_triggers.rs - Database trigger generation
- src/codegen/dashboard/docker.rs - Docker configuration
- src/codegen/dashboard/utils.rs - Shared utilities

## CLI Enhancements

Extended nomnom CLI with new commands in src/bin/nomnom.rs:
- `generate-ingestion-api` - Generate Axum ingestion server
- `generate-dashboard` - Generate real-time monitoring dashboard

Both commands support:
- `--entities` - Path to entity YAML directory
- `--output` - Output directory for generated code
- `--database` - Database type (postgresql/mysql/mariadb)
- `--backend` - Backend framework (axum/fastapi for dashboard)

## Core Changes

Parser binary generation (src/codegen/parser_binary.rs):
- Enhanced to generate ingestion API binaries alongside parser binaries
- Improved type mapping and validation
- Better handling of persistent vs transient entities

Entity processing improvements:
- src/codegen/types.rs - New PersistentEntity type
- src/codegen/project_config.rs - Database config validation
- src/codegen/orchestration.rs - Ingestion API orchestration

## Testing Infrastructure

New test scripts in config/examples/tpch/:
- test-api.sh - End-to-end testing of ingestion API
  * Auto-generates server
  * Runs database migrations
  * Tests all ingestion endpoints
  * Validates data persistence

- test-dashboard.sh - End-to-end testing of dashboard
  * Auto-generates and builds dashboard
  * Configures database connection
  * Starts backend (Axum) and frontend (Vite)
  * Automatic cleanup on exit
  * Support for --debug and --clean modes

- generate.sh - Unified script to regenerate all components
  * Parser binaries
  * Ingestion API server
  * Real-time dashboard

## Documentation

Added comprehensive design documentation:
- DESIGN_RUST_INGESTION_API.md - Ingestion API architecture and design
- DESIGN_AXUM_DASHBOARD.md - Dashboard implementation details
- IMPLEMENTATION_STATUS.md - Feature tracking and roadmap
- CLAUDE.md - Development guidelines

## TPCH Example Updates

Enhanced config/examples/tpch with:
- Updated entity YAML files with persistence configurations
- Generated ingestion API server
- Generated real-time dashboard
- Updated README with usage instructions
- Database test scripts and fixtures

## Key Architectural Decisions

1. **Axum over FastAPI**: Chose Rust/Axum for both ingestion API and
   dashboard backend for type safety, performance, and consistency

2. **Direct Table Polling**: Dashboard polls entity tables directly rather
   than using database triggers and event tables, simplifying deployment

3. **Initial Data Loading**: Dashboard sends existing records to clients
   on WebSocket connection, ensuring immediate data visibility

4. **Code Generation Focus**: All components are fully code-generated from
   entity YAML, enabling rapid iteration and consistency

## Breaking Changes

None - all new functionality is opt-in via new CLI commands.

## Testing

- Manual testing with TPC-H example database
- Verified ingestion API endpoints (POST, streaming)
- Confirmed dashboard real-time updates
- Tested with PostgreSQL (primary target)

+++ src/codegen/diesel/schema.rs
@@ -117,7 +117,7 @@ pub fn generate_schema(
                                     }
                                 },
                                 "Integer" => "Integer".to_string(),
-                                "Float" => "Float".to_string(),
+                                "Float" => "Numeric".to_string(),  // Use Numeric for BigDecimal compatibility
                                 "Boolean" => "Bool".to_string(),
                                 "DateTime" => "Timestamp".to_string(),
                                 _ => "Text".to_string(),
